backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
decorator==5.1.1
fastapi==0.89.1
flatbuffers==23.1.4
frozenlist==1.3.3
fsspec==2022.11.0
gcsfs==2022.11.0
-f https://download.pytorch.org/whl/torch_stable.html
albumentations==1.3.0
anyio==3.6.2
cachetools==5.2.1
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
fastapi==0.89.1
flatbuffers==23.1.4
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.24.0
joblib==1.2.0
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
onnxruntime==1.13.1
opencv-python-headless==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2022.10.10
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
